requirements.txt
pygithub
python-dotenv
anthropic
